Misc fixes, 2022 prep
