+++ Aoc/Aoc.Tests/Problems/Aoc21/Problem10Tests.cs
@@ -24,7 +24,7 @@ public class Problem10Tests : ProblemTests<Problem10>
 
     protected override object CorrectOutput1 => 26397;
 
-    protected override object CorrectOutput2 => 288957;
+    protected override object CorrectOutput2 => 288957L;
 
     [TestMethod]
     public void Part1_SmallInput_Is_Correct() => RunPart1(CorrectOutput1, SmallInput);
