+++ Aoc/Aoc.Tests/Problems/Aoc21/Problem3Tests.cs
@@ -16,7 +16,7 @@ public class Problem3Tests
     [DataRow(198, "00100;11110;10110;10111;10101;01111;00111;11100;10000;11001;00010;01010")]
     public void Part1_SmallInput_Is_Correct(object correctOutput, string input)
     {
-        var binaryList = ProblemInput.ParseList(input);
+        var binaryList = InputReader.ParseList(input);
 
         var output = Problem3.MeasurePower(binaryList);
 
@@ -27,7 +27,7 @@ public class Problem3Tests
     [DataRow(230, "00100;11110;10110;10111;10101;01111;00111;11100;10000;11001;00010;01010")]
     public void Part2_SmallInput_Is_Correct(object correctOutput, string input)
     {
-        var binaryList = ProblemInput.ParseList(input);
+        var binaryList = InputReader.ParseList(input);
 
         var output = Problem3.LifeSupportRating(binaryList);
