+++ Aoc/Aoc.Tests/Utils/NumberUtilsTests.cs
@@ -1,4 +1,4 @@
-﻿namespace Aoc19.Tests.Utils
+﻿namespace Aoc.Tests.Utils
 {
     using Aoc.Utils;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
