--- Aoc/Aoc/Utils/ProblemInput.cs
@@ -1,41 +1,0 @@
-﻿namespace Aoc.Utils
-{
-    using System;
-    using System.IO;
-    using System.Linq;
-
-    public static class ProblemInput
-    {
-        public static string ReadInput(int year, int number)
-        {
-            var fileName = $"{year}.{number}.txt";
-            var allInputs = Directory.GetFiles("Inputs", "*.txt", SearchOption.AllDirectories);
-
-            var filePath = allInputs.FirstOrDefault(f => f.EndsWith("\\" + fileName));
-
-            if (File.Exists(filePath))
-            {
-                return File.ReadAllText(filePath);
-            }
-            throw new FileNotFoundException($"Input file {fileName} could not be found.", fileName);
-        }
-
-        public static string[] ParseList(string input, params string[] separators)
-        {
-            if (separators.Length > 0)
-            {
-                return input.Split(separators, StringSplitOptions.RemoveEmptyEntries);
-            }
-            else
-            {
-                return input.Split(new string[] { Environment.NewLine, ";" }, StringSplitOptions.RemoveEmptyEntries);
-            }
-
-        }
-
-        public static int[] ParseNumberList(string input, params string[] separators)
-        {
-            return ParseList(input, separators).Select(int.Parse).ToArray();
-        }
-    }
-}
