+++ Aoc/Aoc/Utils/StringUtils.cs
